chore: rename `Scanner` to `AstScanner`
