+++ crates/rolldown/src/bundler/ast_scanner/mod.rs
